requirements.txt
spotipy>=2.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
umap-learn>=0.5.3
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
jupyter>=1.0.0
scipy>=1.11.0
networkx>=3.1
tqdm>=4.65.0